a smallint is an i16, and use .filter instead of .find

+++ src/models/mod.rs
@@ -12,6 +12,6 @@ pub struct NewPage {
 #[derive(Queryable, Serialize)]
 pub struct Page {
     pub id: Uuid,
-    pub item_type: i32,
+    pub item_type: i16,
     pub title: String,
 }
